requirements.txt
# Core dependencies
pandas>=2.0.0         # DataFrame, CSV, Excel handling
numpy>=2.0.0          # Numeric operations, DataFrame support
openpyxl>=3.1.0       # Excel file reading/writing
xlsxwriter>=3.0.0     # Excel file writing/formatting
python-dotenv>=1.0.0  # .env file support
openai>=1.0.0         # OpenAI API client
PyPDF2>=3.0.0         # PDF parsing (abstracts)
PyMuPDF>=1.22.0       # PDF parsing (full publications)
SQLAlchemy>=2.0.0     # Database ORM
structlog>=23.1.0     # Structured logging
rich>=13.0.0          # Rich logging output
tiktoken>=0.5.0       # Token counting for OpenAI
requests>=2.31.0      # HTTP client for ClinicalTrials.gov API

# PDF Processing - Marker Integration
marker>=0.0.1         # Advanced PDF processing with AI
torch>=2.0.0          # PyTorch for Marker's deep learning models
transformers>=4.30.0  # Hugging Face transformers for Marker

# Type checking
typing-extensions>=4.5.0
mypy>=1.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Development tools
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0 